+++ firestore-bigquery-export/functions/__tests__/jest.setup.ts
@@ -1,1 +1,2 @@
+process.env.CI_TEST = "true";
 global.config = () => require("../src/config").default;