+++ firestore-bigquery-export/functions/src/util.ts
@@ -14,9 +14,6 @@
  * limitations under the License.
  */
 
-import { DocumentSnapshot } from "firebase-functions/lib/v1/providers/firestore";
-import { Change } from "firebase-functions";
-
 import { ChangeType } from "./change-tracker";
 
 /**
@@ -24,7 +21,7 @@ import { ChangeType } from "./change-tracker";
  * @param change Firestore document change object.
  * @returns {ChangeType} The type of change.
  */
-export function getChangeType(change: Change<DocumentSnapshot>): ChangeType {
+export function getChangeType(change: { before: any; after: any }): ChangeType {
   if (!change.after.exists) {
     return ChangeType.DELETE;
   }
@@ -39,7 +36,7 @@ export function getChangeType(change: Change<DocumentSnapshot>): ChangeType {
  * @param change Firestore document change object.
  * @returns {string} The document ID.
  */
-export function getDocumentId(change: Change<DocumentSnapshot>): string {
+export function getDocumentId(change: { before: any; after: any }): string {
   if (change.after.exists) {
     return change.after.id;
   }